fields added

+++ routes/product.routes.js
@@ -3,11 +3,16 @@ const { isAuthenticated } = require("../middleware/jwt.middleware");
 const router = express.Router();
 const Product = require("../models/Product.model");
 const Store = require("../models/Store.model")
+const mongoose = require("mongoose")
 
 
 //all products from 1 store
 router.get("/:idStore/products", async (req, res, next) => {
   const { idStore } = req.params;
+
+  if(!mongoose.Types.ObjectId.isValid(idStore)){
+    return;
+  }
   try {
     const products = await Product.find({ _store: idStore });
     res.status(200).json({ result: products });
@@ -18,12 +23,15 @@ router.get("/:idStore/products", async (req, res, next) => {
   }
 });
 
-router.post(
-  "/:idStore/createProduct",
+router.post( "/:idStore/createProduct",
   isAuthenticated,
   async (req, res, next) => {
     const { idStore } = req.params;
     const {_id} = req.payload
+    if(!mongoose.Types.ObjectId.isValid(idStore)){
+        return;
+      }
+
     try {
         const isMyStore = await Store.findOne({_id:idStore, _owner:_id })
         if(!isMyStore){
@@ -41,6 +49,10 @@ router.patch("/:id/edit", isAuthenticated, async (req, res, next) => {
   const { id } = req.params;
   const { name, description } = req.body;
   const { _id } = req.payload;
+
+  if(!mongoose.Types.ObjectId.isValid(id)){
+    return;
+  }
   //add midlewate p images
   try {
 
@@ -57,6 +69,10 @@ router.patch("/:id/edit", isAuthenticated, async (req, res, next) => {
 
 router.delete("/:id/delete", isAuthenticated, async (req, res, next) => {
   const { id } = req.params;
+
+  if(!mongoose.Types.ObjectId.isValid(id)){
+    return;
+  }
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ message: "product deleted" });
