+++ routes/review.routes.js
@@ -2,11 +2,15 @@ const express = require("express");
 const { isAuthenticated } = require("../middleware/jwt.middleware");
 const router = express.Router();
 const Review = require("../models/Review.model");
+const mongoose = require("mongoose")
 
 
 //get all reviews from 1 store
 router.get("/:storeId/reviews", isAuthenticated, async(req,res,next)=>{
     const {storeId} = req.params
+    if(!mongoose.Types.ObjectId.isValid(storeId)){
+      return;
+    }
     try{
         const allRevStore = await Review.find({_store:storeId})
         res.status(200).json({result: allRevStore})
@@ -19,6 +23,9 @@ router.get("/:storeId/reviews", isAuthenticated, async(req,res,next)=>{
 router.post("/:storeId/newReview", isAuthenticated, async (req, res, next) => {
     const {_id} =req.payload
     const {storeId} = req.params
+    if(!mongoose.Types.ObjectId.isValid(storeId)){
+      return;
+    }
   try {
     const newReview = await Review.create({...req.body, _store:storeId, _owner: _id});
     res.status(201).json({result: newReview});
@@ -31,6 +38,10 @@ router.patch("/:idReview/edit", isAuthenticated,  async (req, res, next) => {
   const { idReview } = req.params;
   const { _owner, _store, ...restBody } = req.body;
   const {_id} = req.payload
+
+  if(!mongoose.Types.ObjectId.isValid(idReview)){
+    return;
+  }
   try {
     const reviewUpdate = await Review.findOneAndUpdate({_id:idReview, _owner:_id}, restBody, {new: true});
     res.status(200).json({ result: reviewUpdate });
@@ -42,6 +53,10 @@ router.patch("/:idReview/edit", isAuthenticated,  async (req, res, next) => {
 router.delete("/:idReview/delete",isAuthenticated, async(req,res,next)=>{
 const {idReview} =req.params
 const {_id}= req.payload
+
+if(!mongoose.Types.ObjectId.isValid(idReview)){
+  return;
+}
     try{
 await Review.findOneAndDelete({_id:idReview, _owner:_id});
 res.status(200).json({ message: "Review deleted" })
